Fix #125: Use only selected image
I'm not 100% sure what the implications are for the media import, but it
appears that only radios will ever be used in this JS, and thus the
array should only contain that one element. Prior to this commit, every
radio change resulted in an array push, and thus #125 would occur, where
the first radio selection would be passed to the creator, and the
intended one would be the last in the array.