Add a loop to prevent hash collisions
Here's my attempt at making something that prevents hash collisions.

It's a simple loop that will keep going until it succeeds or is killed
by the PHP engine for running too long. I'm not sure what the failure
case for this should be. In theory, if the loop ran infinitely, it would
mean we have hit a point where it is near impossible to find an unused
instance ID. This would mean all kinds of other things are broken, and
re-running the request from the client would most likely not solve it.

But anyway, I'm not sure what the fail condition would be, other than dying from execution time.

Updates #451 Instances can create duplicate hashes and fail badly