[fix] notification email from system user working
Changes in 7.0.0 changed how object permissions expirations were
handled.  It now purges all old permissions and notifies before checking
for access.  This require notifications to be sent by the server instead
of a specific user. Due to an oversight, lack of testing, and a dev env
that has email disabled - the code that actually sends the mail did not
receive associated updates that it needed.

This solves the problem by extracting the code that resolves the from
user for notifications so it can be used in both places.

This code also changes notifications so that when one is sent, no longer
are all pending notifications sent.  This change was made to prevent old
notifications that have gone unsent from potentially causing errors in
unrelated code (if a bug occurs in sending a notification, it will keep
occuring - blocking other notifications).

This means we no longer have code that will send past failed notifications.