Upgrades to php 8.0.24 and fuelphp 1.9-develop
PHP support for 7 is ending very soon, so this is
and attempt to update to give the team some more runway.
PHP 8.0.x doesn't have a very long life either, but this
is better than nothing.

Upgrading to PHP 8.0.x required upgrading fuel to a development
version unfortunately - there's no official full release that has
the updates required to work with PHP 8.0.x